Update aria-usage.js
Added aria-palceholder as a supported value for input type="number".

Added check for aria-placeholder being used with the placeholder attribute.

Updated rules to allow aria-placeholder on input elements with type "date", "month", "week", "time", and "datetime-local"

Added rule to check for aria-readonly. Flags an error if used with a value of "false" along with a native readonly attribute. Otherwise flags a warning that aria-readonly is used with a native readonly attribute.

Added rule for rowspan and aria-rowspan to check the values are the same and provide a warning, otherwise return an error.

Added rule for rowspan and aria-rowspan to check the values are the same and provide a warning, otherwise return an error.

Fixed an uncaught exception error that was causing problems on some pages.