Update aria-usage.js
The unknown element test was preventing real issues being reported. I've reworked the logic so that unknown elements are treated like semantically neutral div elements. I'm not sure of the value of reporting unknown elements, as it isn't really a problem, and we come across it a lot in audits. I've left it in for now.