Allow aria-checked on options
The ARIA practices document indicates that either aria-selected or aria-checked attributes may be used on elements with role="option". If the selection state is indicated with aria-selected, then aria-checked is not specified for any options; if the selection state is indicated with aria-checked, then aria-selected is not specified for any options. Added logic to allow one or other technique to be used on listboxes. Thanks to Will Slone for pointing it out.